Major refactoring of Address class and subclasses
The Linear property was eliminated, and each subclass now implements a ToLinear() method that performs appropriate conversion to linear addresses.

Added MakeAddressFromConstant and MakeAddressFromLinear to the Platform class. Anytime an address is created, it should most likely be done through these two methods.

Added a new project, Ps3.csproj, for Playstation3 binaries.

+++ src/Arch/Arm/ThumbDisassembler.cs
@@ -54,7 +54,7 @@ namespace Decompiler.Arch.Arm
                 Address = rdr.Address,
             };
             Disassemble();
-            thumb.Length = rdr.Address - thumb.Address;
+            thumb.Length = (int)(rdr.Address - thumb.Address);
             return thumb;
         }
